tgui fixes
