+++ tgui/package.json
@@ -4,8 +4,8 @@
 	"type": "module",
 	"devDependencies": {
 		"@happy-dom/global-registrator": "^20.10.6",
-		"@rspack/cli": "^2.1.2",
-		"@rspack/core": "^2.1.2",
+		"@rspack/cli": "^2.1.3",
+		"@rspack/core": "^2.1.3",
 		"@testing-library/react": "^16.3.2",
 		"@types/bun": "^1.3.14",
 		"@types/react": "^19.2.17",
@@ -15,7 +15,7 @@
 		"css-loader": "^7.1.4",
 		"sass-embedded": "^1.100.0",
 		"sass-loader": "^17.0.0",
-		"typescript": "^6.0.3"
+		"typescript": "^7.0.2"
 	},
 	"packageManager": "bun@1.3.3",
 	"private": true,
