+++ tgui/packages/tgui-panel/package.json
@@ -3,7 +3,7 @@
 	"version": "6.0.0",
 	"dependencies": {
 		"common": "workspace:*",
-		"dompurify": "^3.4.11",
+		"dompurify": "^3.4.12",
 		"es-toolkit": "^1.49.0",
 		"jotai": "^2.20.1",
 		"react": "^19.2.7",
