+++ tgui/packages/tgui-setup/package.json
@@ -9,6 +9,6 @@
 	},
 	"dependencies": {
 		"clean-css-cli": "^5.6.3",
-		"terser": "^5.48.0"
+		"terser": "^5.49.0"
 	}
 }
