+++ tgui/packages/tgui/interfaces/common/ChemicalContents.tsx
@@ -1,12 +1,18 @@
 import { Box, Stack } from 'tgui-core/components';
+import type { Reagent } from '../ChemDispenser/types';
 
-export const formatUnits = (a) => `${a} unit${a === 1 ? '' : 's'}`;
+export function formatUnits(a: number) {
+  return `${a} unit${a === 1 ? '' : 's'}`;
+}
 
 /**
  * Displays a beaker's contents
- * @property {object} props
  */
-export const BeakerContents = (props) => {
+export const BeakerContents = (props: {
+  beakerLoaded: boolean;
+  beakerContents: Reagent[];
+  buttons: (chemical: Reagent, i: number) => React.JSX.Element;
+}) => {
   const { beakerLoaded, beakerContents = [], buttons } = props;
   return (
     <Box>
@@ -15,36 +21,53 @@ export const BeakerContents = (props) => {
           <Box color="label">The beaker is empty.</Box>
         ))}
       {beakerContents.map((chemical, i) => (
-        <Box key={i} width="100%">
-          <Stack align="center" justify="space-between">
-            <Stack.Item color="label">
-              {formatUnits(chemical.volume)} of {chemical.name}
-            </Stack.Item>
-            {!!buttons && <Stack.Item>{buttons(chemical, i)}</Stack.Item>}
-          </Stack>
-        </Box>
+        <ChemEntry
+          key={chemical.name}
+          chemical={chemical}
+          buttons={buttons}
+          index={i}
+        />
       ))}
     </Box>
   );
 };
-/** Like above, but for buffer contents */
-export const BufferContents = (props) => {
+
+export const BufferContents = (props: {
+  bufferContents: Reagent[];
+  buttons: (chemical: Reagent, i: number) => React.JSX.Element;
+}) => {
   const { bufferContents = [], buttons } = props;
   return (
     <Box>
-      {(bufferContents.length === 0 && (
-          <Box color="label">The buffer is empty.</Box>
-        ))}
+      {bufferContents.length === 0 && (
+        <Box color="label">The buffer is empty.</Box>
+      )}
       {bufferContents.map((chemical, i) => (
-        <Box key={i} width="100%">
-          <Stack align="center" justify="space-between">
-            <Stack.Item color="label">
-              {formatUnits(chemical.volume)} of {chemical.name}
-            </Stack.Item>
-            {!!buttons && <Stack.Item>{buttons(chemical, i)}</Stack.Item>}
-          </Stack>
-        </Box>
+        <ChemEntry
+          key={chemical.name}
+          chemical={chemical}
+          buttons={buttons}
+          index={i}
+        />
       ))}
+    </Box>
+  );
+};
+
+const ChemEntry = (props: {
+  chemical: Reagent;
+  buttons: (chemical: Reagent, i: number) => React.JSX.Element;
+  index: number;
+}) => {
+  const { chemical, buttons, index } = props;
+  return (
+    <Box width="100%">
+      <Stack align="center" justify="space-between">
+        <Stack.Item color="label">
+          {formatUnits(chemical.volume)} of {chemical.name}
+        </Stack.Item>
+        {!!buttons && <Stack.Item>{buttons(chemical, index)}</Stack.Item>}
+      </Stack>
     </Box>
   );
 };
