+++ tgui/packages/tgui/package.json
@@ -4,13 +4,13 @@
 	"dependencies": {
 		"common": "workspace:*",
 		"dateformat": "^5.0.3",
-		"dompurify": "^3.4.11",
+		"dompurify": "^3.4.12",
 		"es-toolkit": "^1.49.0",
 		"highlight.js": "^11.11.1",
 		"jotai": "^2.20.1",
 		"js-yaml": "^5.2.1",
-		"marked": "^18.0.5",
-		"marked-base-url": "^1.1.9",
+		"marked": "^18.0.6",
+		"marked-base-url": "^1.1.10",
 		"marked-smartypants": "^1.1.12",
 		"react": "^19.2.7",
 		"react-dom": "^19.2.7",
